// packages/api/src/routes/spells/types.ts
export type SpellsQuerystring = {
  id?: string;
  slug?: string;
  name?: string;
  incantation?: string;
  category?: string;
  effect?: string;
  light?: string;
  hand?: string;
  creator?: string;
  image?: string;
  wiki?: string;
  sort?: string;
  order?: string;
  page?: number;
  size?: number;
}